+++ src/irmin-layers/irmin_layers_intf.ml
@@ -20,31 +20,31 @@ module type S = sig
   include Irmin.S
 
   val freeze :
-    ?min:commit list ->
-    ?max:commit list ->
-    ?squash:bool ->
-    ?copy_in_upper:bool ->
+    ?min_lower:commit list ->
+    ?max_lower:commit list ->
     ?min_upper:commit list ->
+    ?max_upper:commit list ->
     ?recovery:bool ->
     repo ->
     unit Lwt.t
-  (** [freeze ?min ?max ?squash ?copy_in_upper ?min_upper ?recovery t] launches
+  (** [freeze ?min_lower ?max_lower ?min_upper ?max_upper ?recovery t] launches
       an asynchronous freezing operation on the repo [t] to reduce the size of
       the upper layer and discard unnecessary branches of objects (i.e. commits,
       nodes and contents).
 
-      Let [o] be the set of objects reachable from the [max] commits and bounded
-      by the [min] commits. During the freeze, all objects in [o] are copied to
-      the lower layer, if there is one. Setting [squash] to [true] is equivalent
-      to setting [min] to [max]. [min] defaults to the empty list and [max]
-      defaults to the head commits of the repo.
+      Let [o] be the set of objects reachable from the [max_lower] commits and
+      bounded by the [min_lower] commits. During the freeze, all objects in [o]
+      are copied to the lower layer, if there is one. [max_lower] defaults to
+      the head commits of the repo and [min_lower] defaults to the empty list
+      (i.e. the copy is unbounded). When [max_lower] is the empty list, nothing
+      is copied.
 
-      Let [o'] be the set of objects reachable from the [max] commits and
-      bounded by the [min_upper] commits. When the freeze is over, if
-      [copy_in_upper] is true, then the new upper layer will only contain the
-      objects of [o'], otherwise the new upper layer will be empty. [min_upper]
-      defaults to the empty list and [copy_in_upper] defaults to the repo's
-      configuration.
+      Let [o'] be the set of objects reachable from the [max_upper] commits and
+      bounded by the [min_upper] commits. When the freeze is over, the new upper
+      layer will only contain the objects of [o']. [max_upper] defaults to
+      [max_lower] and [min_upper] defaults to [max_upper] (i.e. only the max
+      commits are copied). When [max_upper] is the empty list, nothing is
+      copied.
 
       If [recovery] is true then the function will first try to recover from a
       previously interrupted freeze. See {!needs_recovery}.
@@ -106,11 +106,10 @@ module type S = sig
     val wait_for_freeze : repo -> unit Lwt.t
 
     val freeze' :
-      ?min:commit list ->
-      ?max:commit list ->
-      ?squash:bool ->
-      ?copy_in_upper:bool ->
+      ?min_lower:commit list ->
+      ?max_lower:commit list ->
       ?min_upper:commit list ->
+      ?max_upper:commit list ->
       ?recovery:bool ->
       ?hook:
         [ `After_Clear